refactor: 열거형 분리 및 메서드 이동
* Juice.swift 파일생성 후 Juice 타입 삽입
* FruitStore내 있는 isHaveEnoughStock 메서드를 JuiceMaker내로 이동
