+++ JuiceMaker/JuiceMaker/Model/Juice.swift
@@ -1,0 +1,58 @@
+//
+//  Juice.swift
+//  JuiceMaker
+//
+//  Created by Quokkaaa
+//
+
+import Foundation
+
+enum Juice: String {
+    typealias JuiceIngredient = Int
+    
+    case strawberryJuice
+    case bananaJuice
+    case kiwiJuice
+    case pineappleJuice
+    case strawberryBananaJuice
+    case mangoJuice
+    case mangoKiwiJuice
+    
+    var recipe: [FruitStore.Fruit: JuiceIngredient] {
+        switch self {
+        case .strawberryJuice:
+            return [.strawberry: 16]
+        case .bananaJuice:
+            return [.banana: 2]
+        case .kiwiJuice:
+            return [.kiwi: 3]
+        case .pineappleJuice:
+            return [.pineapple: 2]
+        case .strawberryBananaJuice:
+            return [.strawberry: 9, .banana: 1]
+        case .mangoJuice:
+            return [.mango: 3]
+        case .mangoKiwiJuice:
+            return [.mango: 2, .kiwi: 1]
+        }
+    }
+    
+    var name: String {
+        switch self {
+        case .strawberryJuice:
+            return "딸기"
+        case .bananaJuice:
+            return "바나나"
+        case .kiwiJuice:
+            return "키위"
+        case .pineappleJuice:
+            return "파인애플"
+        case .strawberryBananaJuice:
+            return "딸바"
+        case .mangoJuice:
+            return "망고"
+        case .mangoKiwiJuice:
+            return "망고키위"
+        }
+    }
+}
